feat(babel-preset-udemy-website): Split config based on legacy/modern browsers
BREAKING CHANGE: 'device' config option has been renamed to 'environment', which has valid values 'legacy' and 'modern'
