+++ packages/babel-preset-udemy-website/tests/result/legacy/template.js
@@ -4,11 +4,14 @@ import "core-js/modules/es6.array.copy-within";
 import "core-js/modules/es6.array.fill";
 import "core-js/modules/es6.array.find";
 import "core-js/modules/es6.array.find-index";
+import "core-js/modules/es7.array.flat-map";
 import "core-js/modules/es6.array.from";
 import "core-js/modules/es7.array.includes";
 import "core-js/modules/es6.array.iterator";
 import "core-js/modules/es6.array.of";
+import "core-js/modules/es6.array.sort";
 import "core-js/modules/es6.array.species";
+import "core-js/modules/es6.date.to-json";
 import "core-js/modules/es6.date.to-primitive";
 import "core-js/modules/es6.function.has-instance";
 import "core-js/modules/es6.function.name";
@@ -52,12 +55,14 @@ import "core-js/modules/es6.object.get-prototype-of";
 import "core-js/modules/es7.object.lookup-getter";
 import "core-js/modules/es7.object.lookup-setter";
 import "core-js/modules/es6.object.prevent-extensions";
+import "core-js/modules/es6.object.to-string";
 import "core-js/modules/es6.object.is";
 import "core-js/modules/es6.object.is-frozen";
 import "core-js/modules/es6.object.is-sealed";
 import "core-js/modules/es6.object.is-extensible";
 import "core-js/modules/es6.object.keys";
 import "core-js/modules/es6.object.seal";
+import "core-js/modules/es6.object.set-prototype-of";
 import "core-js/modules/es7.object.values";
 import "core-js/modules/es6.promise";
 import "core-js/modules/es7.promise.finally";
@@ -107,6 +112,8 @@ import "core-js/modules/es6.string.starts-with";
 import "core-js/modules/es6.string.strike";
 import "core-js/modules/es6.string.sub";
 import "core-js/modules/es6.string.sup";
+import "core-js/modules/es7.string.trim-left";
+import "core-js/modules/es7.string.trim-right";
 import "core-js/modules/es6.typed.array-buffer";
 import "core-js/modules/es6.typed.int8-array";
 import "core-js/modules/es6.typed.uint8-array";
@@ -123,33 +130,15 @@ import "core-js/modules/web.timers";
 import "core-js/modules/web.immediate";
 import "core-js/modules/web.dom.iterable";
 import "regenerator-runtime/runtime";
-var array = new Array(1e3).fill(1);
-var counter = 0;
-var _iteratorNormalCompletion = true;
-var _didIteratorError = false;
-var _iteratorError = undefined;
 
-try {
-  for (var _iterator = array.entries()[Symbol.iterator](), _step; !(_iteratorNormalCompletion = (_step = _iterator.next()).done); _iteratorNormalCompletion = true) {
-    var _step$value = babelHelpers.slicedToArray(_step.value, 2),
-        i = _step$value[0],
-        val = _step$value[1];
+function _templateObject() {
+  var data = babelHelpers.taggedTemplateLiteral(["1", "3"]);
 
-    counter += i + val;
-  }
-} catch (err) {
-  _didIteratorError = true;
-  _iteratorError = err;
-} finally {
-  try {
-    if (!_iteratorNormalCompletion && _iterator.return != null) {
-      _iterator.return();
-    }
-  } finally {
-    if (_didIteratorError) {
-      throw _iteratorError;
-    }
-  }
+  _templateObject = function _templateObject() {
+    return data;
+  };
+
+  return data;
 }
 
-export default counter;
+export default String.raw(_templateObject(), 2);
