--- packages/babel-preset-udemy-website/tests/result/mobile/object-spread.js
@@ -1,22 +1,0 @@
-"use strict";
-
-import "core-js/modules/es7.promise.finally";
-import "core-js/modules/es7.symbol.async-iterator";
-import "core-js/modules/web.timers";
-import "core-js/modules/web.immediate";
-import "core-js/modules/web.dom.iterable";
-import "regenerator-runtime/runtime";
-
-function React() {}
-
-function Component() {}
-
-export default [babelHelpers.objectSpread({
-  a: 1
-}, {
-  b: 2
-}), React.createElement(Component, Object.assign({
-  a: 1
-}, {
-  b: 2
-}))];
