+++ packages/babel-preset-udemy-website/tests/result/modern/arrows.js
@@ -1,0 +1,25 @@
+"use strict";
+
+import "core-js/modules/es6.number.is-finite";
+import "core-js/modules/es6.number.is-nan";
+import "core-js/modules/es6.object.is";
+import "core-js/modules/es6.string.anchor";
+import "core-js/modules/es6.string.big";
+import "core-js/modules/es6.string.blink";
+import "core-js/modules/es6.string.bold";
+import "core-js/modules/es6.string.fixed";
+import "core-js/modules/es6.string.fontcolor";
+import "core-js/modules/es6.string.fontsize";
+import "core-js/modules/es6.string.italics";
+import "core-js/modules/es6.string.link";
+import "core-js/modules/es6.string.small";
+import "core-js/modules/es6.string.strike";
+import "core-js/modules/es6.string.sub";
+import "core-js/modules/es6.string.sup";
+import "core-js/modules/web.timers";
+import "core-js/modules/web.immediate";
+import "core-js/modules/web.dom.iterable";
+import "regenerator-runtime/runtime";
+export default (function (it) {
+  return it;
+});
