+++ packages/babel-preset-udemy-website/tests/result/modern/subclassing-jsx-decorators-initializers.js
@@ -1,7 +1,21 @@
 "use strict";
 
-import "core-js/modules/es7.promise.finally";
-import "core-js/modules/es7.symbol.async-iterator";
+import "core-js/modules/es6.number.is-finite";
+import "core-js/modules/es6.number.is-nan";
+import "core-js/modules/es6.object.is";
+import "core-js/modules/es6.string.anchor";
+import "core-js/modules/es6.string.big";
+import "core-js/modules/es6.string.blink";
+import "core-js/modules/es6.string.bold";
+import "core-js/modules/es6.string.fixed";
+import "core-js/modules/es6.string.fontcolor";
+import "core-js/modules/es6.string.fontsize";
+import "core-js/modules/es6.string.italics";
+import "core-js/modules/es6.string.link";
+import "core-js/modules/es6.string.small";
+import "core-js/modules/es6.string.strike";
+import "core-js/modules/es6.string.sub";
+import "core-js/modules/es6.string.sup";
 import "core-js/modules/web.timers";
 import "core-js/modules/web.immediate";
 import "core-js/modules/web.dom.iterable";
@@ -34,7 +48,7 @@ function (_Component) {
     }
 
     _this = babelHelpers.possibleConstructorReturn(this, (_babelHelpers$getProt = babelHelpers.getPrototypeOf(Class)).call.apply(_babelHelpers$getProt, [this].concat(args)));
-    babelHelpers.initializerDefineProperty(_this, "prop1", _descriptor, babelHelpers.assertThisInitialized(babelHelpers.assertThisInitialized(_this)));
+    babelHelpers.initializerDefineProperty(_this, "prop1", _descriptor, babelHelpers.assertThisInitialized(_this));
     _this.prop2 = 14;
     return _this;
   }
